Reorganize module hierarchy
