+++ src/Codec/Compression/LZMA.hs
@@ -1,38 +1,30 @@
 module Codec.Compression.LZMA
-  ( -- * Simple compression and decompression
+  (
+  -- | This module provides pure functions for compressing and decompressing
+  -- streams of data in the xz format and represented by lazy 'L.ByteString's.
+  -- This makes it easy to use either in memory or with disk or network IO.
+
+  -- * Simple compression and decompression
     decompress
 
-  -- * Extended API with control over compression parameters
+  -- * Extended API with control over (de)compression parameters
   , decompressWith
-  , DecompressParams(..)
-  , defaultDecompressParams
+  , I.DecompressParams(..)
+  , I.defaultDecompressParams
 
-  -- * Streaming interface for decompression
-  , DecompressStream
-  , DecompressError(..)
-  , decompressST
-  , decompressIO
-
-  -- * Decompression with random seek support
-  , Index.Index
-  , Index.ReadRequest(..)
-  , IndexedDecompressStream
-  , indexedDecompressIO
   ) where
 import qualified Data.ByteString.Lazy as L
 
-import Codec.Compression.LZMA.Internal hiding (decompress)
-import qualified Codec.Compression.LZMA.Index as Index
-import qualified Codec.Compression.LZMA.Internal as Internal
+import qualified Codec.Compression.LZMA.Internal as I
 
 -- | Decompress a stream of data in the xz format
 decompress :: L.ByteString -> L.ByteString
-decompress = decompressWith defaultDecompressParams
+decompress = decompressWith I.defaultDecompressParams
 
 -- | Like 'decompress' but with the ability to specify various decompression
 -- parameters. Typical usage:
 --
 -- > decompressWith defaultCompressParams { ... }
 --
-decompressWith :: DecompressParams -> L.ByteString -> L.ByteString
-decompressWith = Internal.decompress
+decompressWith :: I.DecompressParams -> L.ByteString -> L.ByteString
+decompressWith = I.decompress
