+++ src/Codec/Compression/LZMA/Internal.hs
@@ -2,11 +2,13 @@
 {-# LANGUAGE CPP #-}
 {-# LANGUAGE DataKinds #-}
 {-# LANGUAGE DeriveDataTypeable #-}
-{-# LANGUAGE FlexibleContexts #-}
+{-# LANGUAGE GADTs #-}
 {-# LANGUAGE NamedFieldPuns #-}
 {-# LANGUAGE RecordWildCards #-}
+{-# LANGUAGE ViewPatterns #-}
 module Codec.Compression.LZMA.Internal
-  ( -- * Decompression
+  (
+  -- * Decompression
     DecompressParams(..)
   , defaultDecompressParams
 
@@ -14,25 +16,39 @@ module Codec.Compression.LZMA.Internal
   , decompress
 
   -- ** Incremental processing
-  -- *** Decompression
+  -- *** Sequential decompression
   , DecompressStream, DecompressError(..)
   , decompressST
   , decompressIO
-  -- , decompressStream
-  -- ** Random access
-  , IndexedDecompressStream
-  , indexedDecompressIO
+  , decompressStream
+  -- *** Decompression with random access support
+  , SeekableDecompressStream
+  , seekableDecompressIO
+  , seekableDecompressStream
+  -- *** Decoding indicies
+  , decodeIndicies
+  , DecodeStream
+  , decodeIndexStream
+  , ID.StreamPadding
+
+  -- * Types for incremental processing
+  , ReadRequest(..)
+  , Position
+  , Compression(..)
+  , Size
   ) where
 import Control.Applicative
-import Control.Exception (assert)
+import Control.Exception (IOException, assert)
 import Control.Monad
 import Data.Maybe (fromMaybe)
 import Data.Monoid
 import Data.Typeable (Typeable)
 import Data.Word
 import Foreign hiding (void)
+import System.IO
 
 import Control.Monad.Catch
+import Control.Monad.Trans
 import Pipes hiding (next, void)
 import Pipes.Core
 import qualified Control.Monad.ST as S
@@ -43,10 +59,12 @@ import qualified Data.ByteString.Lazy as L
 import qualified Data.ByteString.Lazy.Internal as L
 import qualified Pipes.Internal as P
 
-import Codec.Compression.LZMA.Internal.Index
+import Codec.Compression.LZMA.Internal.IndexDecoder (IndexDecoder)
 import Codec.Compression.LZMA.Internal.Stream (Stream)
-import qualified Codec.Compression.LZMA.Internal.Stream as Stream
+import Codec.Compression.LZMA.Internal.Types
 import qualified Codec.Compression.LZMA.Internal.C as C
+import qualified Codec.Compression.LZMA.Internal.IndexDecoder as ID
+import qualified Codec.Compression.LZMA.Internal.Stream as Stream
 
 #if DEBUG
 import Debug.Trace
@@ -71,8 +89,11 @@ data DecompressParams = DecompressParams
   -- be converted to a strict 'S.ByteString' in @O(1)@ time using
   -- @'S.concat' . 'L.toChunks'@.
   , decompressMemoryLimit :: !Word64
+  -- ^
   }
 
+-- | The default set of parameters for decompresssion. This is typically used
+-- with the 'decompressWith' function with specific parameters overridden.
 defaultDecompressParams :: DecompressParams
 defaultDecompressParams = DecompressParams
   { decompressBufferSize = defaultDecompressBufferSize
@@ -216,24 +237,28 @@ decompressStreamToIO stream = do
 
 ------------------------------------------------------------
 
-type IndexedDecompressStream = Proxy
+type SeekableDecompressStream = Proxy
   (ReadRequest 'Compressed) S.ByteString
   (ReadRequest 'Uncompressed) S.ByteString
 
-indexedDecompressIO
+seekableDecompressIO
   :: DecompressParams
-  -> Index
+  -> C.Index
+  -- ^ Index of the stream
   -> ReadRequest 'Uncompressed
-  -> IndexedDecompressStream IO ()
-indexedDecompressIO params index =
-  decompressStreamToIO . indexedDecompressStream params index
+  -- ^ Initial request
+  -> SeekableDecompressStream IO ()
+seekableDecompressIO params index =
+  decompressStreamToIO . seekableDecompressStream params index
 
-indexedDecompressStream
+seekableDecompressStream
   :: DecompressParams
-  -> Index
+  -> C.Index
+  -- ^ Index for the stream
   -> ReadRequest 'Uncompressed
-  -> IndexedDecompressStream Stream ()
-indexedDecompressStream params index req0 = do
+  -- ^ Initial request
+  -> SeekableDecompressStream Stream ()
+seekableDecompressStream params index req0 = do
   iter <- liftIO $ C.lzma_index_iter_init index
   decodeLoop iter req0
   liftIO $ C.touchIndex index
@@ -248,13 +273,13 @@ indexedDecompressStream params index req0 = do
     decodeBlock
       :: C.IndexIterPtr
       -> ReadRequest 'Uncompressed
-      -> IndexedDecompressStream Stream (ReadRequest 'Uncompressed)
+      -> SeekableDecompressStream Stream (ReadRequest 'Uncompressed)
     decodeBlock iter req = do
       lift Stream.flushBuffers
-      indexIter@IndexIter {..} <- liftIO $ withForeignPtr iter peek
+      indexIter@C.IndexIter {..} <- liftIO $ withForeignPtr iter peek
       let
-        IndexIterStream {indexIterStreamBlockCount} = indexIterStream
-        IndexIterBlock {..} = indexIterBlock
+        C.IndexIterStream {indexIterStreamBlockCount} = indexIterStream
+        C.IndexIterBlock {..} = indexIterBlock
         blockPos :: Position 'Compressed
         blockPos = fromIntegral indexIterBlockCompressedFileOffset
         blockUncompPos :: Position 'Uncompressed
@@ -290,7 +315,7 @@ indexedDecompressStream params index req0 = do
       :: C.IndexIterPtr
       -> ReadRequest 'Compressed
       -> Int -- ^ Offset from the beginning of the block to the target position
-      -> IndexedDecompressStream Stream (ReadRequest 'Uncompressed)
+      -> SeekableDecompressStream Stream (ReadRequest 'Uncompressed)
     fillBuffers' iter req skipBytes = do
       isLastChunk <- fillBuffers params req
       drainBuffers iter isLastChunk skipBytes
@@ -299,7 +324,7 @@ indexedDecompressStream params index req0 = do
       :: C.IndexIterPtr
       -> Bool -- ^ Last chunk or not
       -> Int -- ^ Offset from the beginning of the block to the target position
-      -> IndexedDecompressStream Stream (ReadRequest 'Uncompressed)
+      -> SeekableDecompressStream Stream (ReadRequest 'Uncompressed)
     drainBuffers iter isLastChunk skipBytes = do
       lift $ assertBuffers isLastChunk
 
@@ -345,7 +370,7 @@ indexedDecompressStream params index req0 = do
 locateBlock
   :: C.IndexIterPtr
   -> ReadRequest 'Uncompressed
-  -> IndexedDecompressStream Stream Bool
+  -> SeekableDecompressStream Stream Bool
 locateBlock iter req = not <$> liftIO act
   where
     act = case req of
@@ -402,3 +427,236 @@ finalizeStream skipBytes = do
     else do
       lift Stream.end
       return Nothing
+
+-----------------------------------------------------------
+-- Index decoder
+
+decodeIndicies :: Handle -> IO (C.Index, ID.StreamPadding)
+decodeIndicies h = do
+  size <- hFileSize h
+  C.allocaStreamFlags $ \header ->
+    C.allocaStreamFlags $ \footer -> do
+      runDecodeStream h $ indexDecodingToIO
+        (decodeIndexStream (fromIntegral size))
+        ID.newIndexDecoderState header footer
+
+runDecodeStream
+  :: (MonadIO m, MonadThrow m)
+  => Handle
+  -> DecodeStream m a
+  -> m a
+runDecodeStream h = runEffect . loop
+  where
+    loop (P.Request seekRequest next) = do
+      size <- case seekRequest of
+        PReadWithSize (fromIntegral -> pos) size -> do
+          r <- liftIO $ try $ hSeek h AbsoluteSeek pos
+          case r of
+            Left e -> lift $ throwM (e :: IOException)
+            Right () -> return size
+        _ -> return L.defaultChunkSize
+      chunk <- liftIO $ hGetEnsureN h size
+      loop (next chunk)
+    loop (P.Respond _ next) = loop (next ())
+    loop (P.M m) = lift m >>= loop
+    loop (P.Pure a) = return a
+
+hGetEnsureN :: Handle -> Int -> IO S.ByteString
+hGetEnsureN h expectedLen = do
+  fp <- S.mallocByteString expectedLen
+  withForeignPtr fp $ \p -> do
+    actualLen <- hGetBuf h p expectedLen
+    if expectedLen <= actualLen
+      then return $ S.PS fp 0 expectedLen
+      else fail $ "hGetEnsureN: unexpected end of file (demanded " ++
+        show expectedLen ++ " bytes, but got " ++ show actualLen ++
+        " bytes)"
+
+-- | Seek to an absolute position and ask for an input with given size.
+pread
+  :: Size
+  -- ^ Input size
+  -> DecodeStream IndexDecoder S.ByteString
+pread size = do
+  pos <- lift ID.getPosition
+  request $ PReadWithSize pos size
+
+-- | Decode things from compressed stream.
+type DecodeStream = Client (ReadRequest 'Compressed) S.ByteString
+
+
+-- | Seek operation failure in downstream.
+data SeekException = SeekError C.ErrorCode deriving (Show, Typeable)
+
+instance Exception SeekException
+
+------------------------------------------------------------
+
+headerSize, footerSize :: Integral a => a
+headerSize = fromIntegral C.streamHeaderSize
+footerSize = headerSize
+
+-- | Seek thorough a seekable stream and build a combined index. The index can
+-- later be used for seeking.
+decodeIndexStream
+  :: Size -- ^ Size of the file
+  -> DecodeStream IndexDecoder (C.Index, ID.StreamPadding)
+decodeIndexStream fileSize = do
+  lift $ ID.setPosition $ fromIntegral fileSize
+  index <- parseIndex
+  loop index
+  padding <- lift ID.getStreamPadding
+  return (index, padding)
+  where
+    loop :: C.Index -> DecodeStream IndexDecoder ()
+    loop index = do
+      pos <- lift ID.getPosition
+      when (pos > 0) $ do
+        index' <- parseIndex
+        handleRet $ liftIO $ C.lzma_index_cat index index' Nothing
+        loop index
+
+-- | Parse an index
+parseIndex :: DecodeStream IndexDecoder C.Index
+parseIndex = do
+  padding <- decodeStreamFooter
+  index <- decodeIndex 8192 -- FIXME: Set appropreate size
+  decodeStreamHeader index
+  checkIntegrity index
+  handleRet $ liftIO $ C.lzma_index_stream_padding index padding
+  lift $ ID.modifyStreamPadding' (+ padding)
+  return index
+
+decodeStreamFooter :: DecodeStream IndexDecoder ID.StreamPadding
+decodeStreamFooter = loop 0
+  where
+    loop padding = do
+      pos <- lift ID.getPosition
+      when (pos < 2 * headerSize) $
+        lift $ throwM $ SeekError C.DataError
+
+      pos' <- lift $ do
+        ID.modifyPosition' (subtract footerSize)
+        ID.getPosition
+      when (pos' < footerSize) $
+        lift $ throwM $ SeekError C.DataError
+
+      chunk <- pread footerSize
+      if isStreamPadding $ dropStreamPadding 2 chunk
+        then do
+          padding' <- lift $ skipStreamPaddings chunk 2 padding
+          loop padding'
+        else do
+          footer <- lift ID.getStreamFooter
+          handleRet $ liftIO $ do
+            let S.PS inFPtr _off _len = chunk
+            withForeignPtr inFPtr $ C.lzma_stream_footer_decode footer
+          -- TODO: check the version of the stream footer
+          return padding
+
+skipStreamPaddings
+  :: S.ByteString
+  -- ^ Input chunk
+  -> Int
+  -- ^ Index
+  -> ID.StreamPadding
+  -> IndexDecoder ID.StreamPadding
+skipStreamPaddings chunk = go
+  where
+    go !i !padding =
+      if i >= 0 && isStreamPadding (dropStreamPadding i chunk)
+        then do
+          ID.modifyPosition' (subtract 4)
+          go (i - 1) (padding + 4)
+        else return padding
+
+isStreamPadding :: S.ByteString -> Bool
+isStreamPadding = S.all (== 0) . S.take 4
+
+dropStreamPadding :: Int -> S.ByteString -> S.ByteString
+dropStreamPadding n = S.drop (4*n)
+
+getIndexSize :: IndexDecoder C.VLI
+getIndexSize = do
+  footer <- ID.getStreamFooter
+  liftIO $ C.lzma_stream_flags_backward_size footer
+
+-- | Decode a stream index.
+decodeIndex
+  :: C.VLI -- ^ Buffer size
+  -> DecodeStream IndexDecoder C.Index
+decodeIndex bufSize = do
+  indexSize <- lift getIndexSize
+  -- Set posision to the beginning of the index.
+  lift $ ID.modifyPosition' $ subtract $ fromIntegral indexSize
+  stream <- liftIO C.newStream
+  (ret, indexRef) <- liftIO $ C.lzma_index_decoder stream maxBound -- FIXME: Set proper value
+  unless (ret == C.Ok) $ lift $ throwM $ SeekError C.ProgError
+
+  loop stream indexRef indexSize
+
+  liftIO $ do
+    allocator <- C.lzma_get_stream_allocator stream
+    C.peekIndexRef indexRef allocator
+  where
+    loop stream indexRef indexSize = do
+      let inAvail :: Integral a => a
+          inAvail = fromIntegral $ min bufSize indexSize
+      liftIO $ C.lzma_set_stream_avail_in stream inAvail
+      chunk <- pread inAvail
+      lift $ ID.modifyPosition' (+ inAvail)
+      let indexSize' = indexSize - inAvail
+      liftIO $ do
+        let S.PS inFPtr _offset _len = chunk
+        withForeignPtr inFPtr $ C.lzma_set_stream_next_in stream
+      ret <- liftIO $ C.lzma_code stream C.Run
+      case ret of
+        C.Ok -> loop stream indexRef indexSize'
+        C.Error reason -> lift $ throwM $ SeekError reason
+        C.StreamEnd -> do
+          inAvail' <- liftIO $ C.lzma_get_stream_avail_in stream
+          unless (indexSize' == 0 && inAvail' == 0) $
+            lift $ throwM $ SeekError C.DataError
+
+-- | Decode the stream header and check that its stream flags match the stream
+-- footer.
+decodeStreamHeader
+  :: C.Index
+  -> DecodeStream IndexDecoder ()
+decodeStreamHeader index = do
+  indexSize <- lift getIndexSize
+  lift $ ID.modifyPosition' (subtract $ fromIntegral indexSize + headerSize)
+  blocksSize <- liftIO $ C.lzma_index_total_size index
+  lift $ ID.modifyPosition' (subtract $ fromIntegral blocksSize)
+  chunk <- pread headerSize
+  let S.PS inFPtr _off _len = chunk
+  header <- lift ID.getStreamHeader
+  handleRet $ liftIO $ withForeignPtr inFPtr $ C.lzma_stream_header_decode header
+
+checkIntegrity
+  :: C.Index
+  -> DecodeStream IndexDecoder ()
+checkIntegrity index = do
+  header <- lift ID.getStreamHeader
+  footer <- lift ID.getStreamFooter
+  handleRet $ liftIO $ C.lzma_stream_flags_compare header footer
+  handleRet $ liftIO $ C.lzma_index_stream_flags index footer
+  padding <- lift ID.getStreamPadding
+  handleRet $ liftIO $ C.lzma_index_stream_padding index padding
+
+indexDecodingToIO
+  :: DecodeStream IndexDecoder a
+  -> ID.IndexDecoderState
+  -> C.StreamFlags -- ^ Stream header
+  -> C.StreamFlags -- ^ Stream footer
+  -> DecodeStream IO a
+indexDecodingToIO stream0 state0 header footer = go stream0 state0
+  where
+    go (P.Request req next) state =
+      P.Request req $ \chunk -> go (next chunk) state
+    go (P.Respond out next) state =
+      P.Respond out $ \resp -> go (next resp) state
+    go (P.M m) state = do
+      (state', stream') <- liftIO $ ID.runIndexDecoder m state header footer
+      go stream' state'
+    go (P.Pure a) _ =  P.Pure a
